feat: create chat report service
