+++ src/main/kotlin/com/example/chatbot/dto/report/ChatReportResponse.kt
@@ -1,0 +1,7 @@
+package com.example.chatbot.dto.report
+
+data class ChatReportResponse(
+    val question: String,
+    val answer: String,
+    val userEmail: String
+)
